REF: Moved MostRecentAnalysis components to the features/articles folder
